+++ Changelog.md
@@ -6,6 +6,7 @@ All significant changes are listed in this file.
 
 - Fix issue with scraper downloading incorrect movie metadata in rare circumstances
 - Fix issues when editing actresses in movie detail view
+- Fix issue when parsing longer IDs
 
 ## [1.1.8] - 2022-11-22